requirements.txt
anyio==3.6.1
APScheduler==3.9.1
astor==0.8.1
async-timeout==4.0.2
attrs==22.1.0
bandit==1.7.4
bleach==6.0.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cryptography==40.0.1
darglint==1.8.1
dependency-injector==4.40.0
docutils==0.19
dotty-dict==1.3.1
eradicate==2.1.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.9.23
flake8-commas==2.1.0
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
gitdb==4.0.9
GitPython==3.1.29
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==1.1.1
invoke==1.7.3
isort==5.10.1
jaraco.classes==3.2.3
keyring==23.13.1
loguru==0.6.0
mccabe==0.6.1
more-itertools==9.0.0
packaging==21.3
pbr==5.10.0
pep8-naming==0.13.2
pkginfo==1.9.6
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-gitlab==3.13.0
python-semantic-release==7.33.1
python-telegram-bot==20.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
readme-renderer==37.3
redis==4.5.1
requests==2.28.2
requests-toolbelt==0.10.1
restructuredtext-lint==1.4.0
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
stevedore==4.0.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
twine==3.8.0
types-pyOpenSSL==23.1.0.1
types-redis==4.5.3.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.14
webencodings==0.5.1
wemake==0.1.0
wemake-python-styleguide==0.17.0
zipp==3.12.1

SQLAlchemy~=2.0.7